Here logic of project is implemented States added in App component are cards_url, cards, score, high_score, selected & game_over. cards_url is an object which contains all the urls of images cards is an array which contains all the keys of images score is the current score high_score is the high score of a specific sever selected is a object which contains the information of clicked cards gameover is true when a clicked card is again clicked
useEffect callback for updating the highscore

onClickCard function checks if the card is already clicked or not. If its clicked then game ends. If not then it will mark that image selected and updates the score.

onRestart function updates the game_over state to true

Various Imports -
cardsId & cardsUrl - these function will return an array of whole numbers and an object of image urls
reshuffle - this function is a used for reshuffling of cards array
AllCards - this is a component for handling all the cards
Header - Name, current score and highscore will be rendered from here
GameOver - This component will mount when game ends